requirements.txt
streamlit==1.42.0
psycopg2-binary
requests
streamlit_js_eval
authlib>=1.3.2
